requirements.txt
gunicorn
flask
pandas
python-dotenv
requests